Ignore spurious iphb wakeups
If dsme issues iphb wakeup when BackgroundActivity object is not waiting
for one, it would cause transition to Running state that the client
application is possibly not prepared to handle.

Ignore valid iphb wakeups if they are not expected.

Also do not clear the m_waiting flag on recoverable I/O errors.

And do not clear m_started flag in case iphb reconnect is needed
so that wakeup is reprogrammed when connection is established again.

[nemo-keepalive] Ignore spurious iphb wakeups. Fixes JB#24831